update message component
